ATR-431: merge the worter for subtree base and default implementations
