+++ src/Acuminator/Acuminator.Vsix/Acuminator.Vsix.csproj
@@ -260,10 +260,9 @@
     <Compile Include="Tool Windows\CodeMap\ViewModel\Base\INodeWithSymbolItem.cs" />
     <Compile Include="Tool Windows\CodeMap\ViewModel\Sorting\NodeDeclarationOrderComparer.cs" />
     <Compile Include="Tool Windows\CodeMap\ViewModel\Sorting\Sorters\NodesSorter.cs" />
-    <Compile Include="Tool Windows\CodeMap\ViewModel\Sorting\Visitors\CodeMapTreeSorterBase.cs" />
+    <Compile Include="Tool Windows\CodeMap\ViewModel\Sorting\Visitors\CodeMapTreeSorterDefault.cs" />
     <Compile Include="Tool Windows\CodeMap\ViewModel\Sorting\CodeMapSortContext.cs" />
     <Compile Include="Tool Windows\CodeMap\ViewModel\Sorting\Visitors\CodeMapTreeInitialSorter.cs" />
-    <Compile Include="Tool Windows\CodeMap\ViewModel\Sorting\Visitors\CodeMapTreeSorterDefault.cs" />
     <Compile Include="Tool Windows\CodeMap\ViewModel\Sorting\SortType.cs" />
     <Compile Include="Tool Windows\CodeMap\Tooltips\TreeNodeToTooltipConverter.cs" />
     <Compile Include="Tool Windows\CodeMap\Model\DacMemberCategory.cs" />
